implemented xbox-pad-controller for mechs therefore renamed package "controller/keyboard" to "controller/device" tick-methods now receive int instead of long, as the gameloop-itself uses int mechs now take care of ticking their actions themselves, instead of having the controllers tick them calling the jump-action from within the controller now, instead of moving the entity around via the move-method

+++ src/util/Countdown.java
@@ -53,7 +53,7 @@ public class Countdown {
 	 * @param delta
 	 *            the milliseconds to substract from the remaining time
 	 */
-	public void tick(final long delta) {
+	public void tick(final int delta) {
 		_current = Math.max(_current - delta, 0);
 	}
